Drop support for Python 3.3
Python 3.3 can't be obtained via Micromamba from conda-forge, so
it'd be very problematic to handle future support for this version.
It's very old version of Python, so it shouldn't be such a problem.
It could work anyway, but automatic tests after changes won't test
this version, so it could break any time.